Re-organize classes
